Move `onSettingsPage` example to Preinstalled example snap
