+++ packages/examples/packages/preinstalled/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './dialog';
 export * from './result';
+export * from './settings';
