--- packages/examples/packages/settings-page/.depcheckrc.json
@@ -1,18 +1,0 @@
-{
-  "ignore-patterns": ["dist", "coverage"],
-  "ignores": [
-    "@lavamoat/allow-scripts",
-    "@lavamoat/preinstall-always-fail",
-    "@metamask/auto-changelog",
-    "@metamask/eslint-*",
-    "@types/*",
-    "@typescript-eslint/*",
-    "eslint-config-*",
-    "eslint-plugin-*",
-    "jest-silent-reporter",
-    "prettier-plugin-packagejson",
-    "ts-node",
-    "typedoc",
-    "typescript"
-  ]
-}
