--- packages/examples/packages/settings-page/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  extends: ['../../.eslintrc.js'],
-
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-  },
-};
