--- packages/examples/packages/settings-page/package.json
@@ -1,86 +1,0 @@
-{
-  "name": "@metamask/settings-page-example-snap",
-  "version": "0.0.0",
-  "description": "MetaMask example snap demonstrating the use of settings pages",
-  "keywords": [
-    "MetaMask",
-    "Snaps",
-    "Ethereum"
-  ],
-  "homepage": "https://github.com/MetaMask/snaps/tree/main/packages/examples/packages/settings-page#readme",
-  "bugs": {
-    "url": "https://github.com/MetaMask/snaps/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/MetaMask/snaps.git"
-  },
-  "license": "(MIT-0 OR Apache-2.0)",
-  "sideEffects": false,
-  "main": "./dist/bundle.js",
-  "files": [
-    "dist",
-    "snap.manifest.json"
-  ],
-  "scripts": {
-    "build": "mm-snap build",
-    "build:clean": "yarn clean && yarn build",
-    "changelog:update": "../../../../scripts/update-changelog.sh @metamask/settings-page-example-snap",
-    "changelog:validate": "../../../../scripts/validate-changelog.sh @metamask/settings-page-example-snap",
-    "clean": "rimraf \"dist\"",
-    "lint": "yarn lint:eslint && yarn lint:misc --check && yarn changelog:validate && yarn lint:dependencies",
-    "lint:ci": "yarn lint",
-    "lint:dependencies": "depcheck",
-    "lint:eslint": "eslint . --cache --ext js,ts,jsx,tsx",
-    "lint:fix": "yarn lint:eslint --fix && yarn lint:misc --write",
-    "lint:misc": "prettier --no-error-on-unmatched-pattern --loglevel warn \"**/*.json\" \"**/*.md\" \"**/*.html\" \"!CHANGELOG.md\" \"!snap.manifest.json\" --ignore-path ../../../../.gitignore",
-    "publish:preview": "yarn npm publish --tag preview",
-    "since-latest-release": "../../../../scripts/since-latest-release.sh",
-    "start": "mm-snap watch",
-    "test": "jest --reporters=jest-silent-reporter",
-    "test:clean": "jest --clearCache",
-    "test:verbose": "jest --verbose",
-    "test:watch": "jest --watch"
-  },
-  "dependencies": {
-    "@metamask/snaps-sdk": "workspace:^"
-  },
-  "devDependencies": {
-    "@jest/globals": "^29.5.0",
-    "@lavamoat/allow-scripts": "^3.0.4",
-    "@metamask/auto-changelog": "^3.4.4",
-    "@metamask/eslint-config": "^12.1.0",
-    "@metamask/eslint-config-jest": "^12.1.0",
-    "@metamask/eslint-config-nodejs": "^12.1.0",
-    "@metamask/eslint-config-typescript": "^12.1.0",
-    "@metamask/snaps-cli": "workspace:^",
-    "@metamask/snaps-jest": "workspace:^",
-    "@swc/core": "1.3.78",
-    "@swc/jest": "^0.2.26",
-    "@typescript-eslint/eslint-plugin": "^5.42.1",
-    "@typescript-eslint/parser": "^6.21.0",
-    "deepmerge": "^4.2.2",
-    "depcheck": "^1.4.7",
-    "eslint": "^8.27.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.1.5",
-    "eslint-plugin-jsdoc": "^41.1.2",
-    "eslint-plugin-n": "^15.7.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "jest": "^29.0.2",
-    "jest-silent-reporter": "^0.6.0",
-    "prettier": "^2.8.8",
-    "prettier-plugin-packagejson": "^2.5.2",
-    "ts-node": "^10.9.1",
-    "typescript": "~5.3.3"
-  },
-  "engines": {
-    "node": "^18.16 || >=20"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  }
-}
