--- packages/examples/packages/settings-page/snap.config.ts
@@ -1,17 +1,0 @@
-import type { SnapConfig } from '@metamask/snaps-cli';
-import { resolve } from 'path';
-
-const config: SnapConfig = {
-  input: resolve(__dirname, 'src/index.tsx'),
-  server: {
-    port: 8071,
-  },
-  typescript: {
-    enabled: true,
-  },
-  stats: {
-    buffer: false,
-  },
-};
-
-export default config;
