--- packages/examples/packages/settings-page/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "composite": false,
-    "baseUrl": "./"
-  },
-  "include": ["src", "snap.config.ts"]
-}
