+++ packages/test-snaps/package.json
@@ -66,7 +66,6 @@
     "@metamask/notification-example-snap": "workspace:^",
     "@metamask/preinstalled-example-snap": "workspace:^",
     "@metamask/send-flow-example-snap": "workspace:^",
-    "@metamask/settings-page-example-snap": "workspace:^",
     "@metamask/signature-insights-example-snap": "workspace:^",
     "@metamask/snaps-utils": "workspace:^",
     "@metamask/utils": "^10.0.0",
