+++ packages/test-snaps/src/features/snaps/index.ts
@@ -27,4 +27,3 @@ export * from './signature-insights';
 export * from './updates';
 export * from './wasm';
 export * from './send-flow';
-export * from './settings-page';
