+++ packages/test-snaps/src/features/snaps/preinstalled/Preinstalled.tsx
@@ -1,6 +1,6 @@
 import { logError } from '@metamask/snaps-utils';
 import type { FunctionComponent } from 'react';
-import { Button } from 'react-bootstrap';
+import { Button, ButtonGroup } from 'react-bootstrap';
 
 import { useInvokeMutation } from '../../../api';
 import { Result, Snap } from '../../../components';
@@ -9,10 +9,17 @@ import { PREINSTALLED_SNAP_ID, PREINSTALLED_VERSION } from './constants';
 export const Preinstalled: FunctionComponent = () => {
   const [invokeSnap, { isLoading, data, error }] = useInvokeMutation();
 
-  const handleSubmit = () => {
+  const handleSubmitDialog = () => {
     invokeSnap({
       snapId: PREINSTALLED_SNAP_ID,
       method: 'showDialog',
+    }).catch(logError);
+  };
+
+  const handleSubmitSettings = () => {
+    invokeSnap({
+      snapId: PREINSTALLED_SNAP_ID,
+      method: 'getSettings',
     }).catch(logError);
   };
 
@@ -23,15 +30,24 @@ export const Preinstalled: FunctionComponent = () => {
       version={PREINSTALLED_VERSION}
       testId="preinstalled-snap"
     >
-      <Button
-        variant="primary"
-        id="showPreinstalledDialog"
-        className="mb-3"
-        disabled={isLoading}
-        onClick={handleSubmit}
-      >
-        Show dialog
-      </Button>
+      <ButtonGroup className="mb-3">
+        <Button
+          variant="primary"
+          id="showPreinstalledDialog"
+          disabled={isLoading}
+          onClick={handleSubmitDialog}
+        >
+          Show dialog
+        </Button>
+        <Button
+          variant="primary"
+          id="settings-state"
+          disabled={isLoading}
+          onClick={handleSubmitSettings}
+        >
+          Get settings state
+        </Button>
+      </ButtonGroup>
       <Result>
         <span id="rpcResult">
           {JSON.stringify(data, null, 2)}
