--- packages/test-snaps/src/features/snaps/settings-page/constants.ts
@@ -1,5 +1,0 @@
-import packageJson from '@metamask/settings-page-example-snap/package.json';
-
-export const SETTINGS_PAGE_SNAP_ID = 'npm:@metamask/settings-page-example-snap';
-export const SETTINGS_PAGE_SNAP_PORT = 8071;
-export const SETTINGS_PAGE_VERSION = packageJson.version;
