--- packages/test-snaps/src/features/snaps/settings-page/index.ts
@@ -1,1 +1,0 @@
-export * from './SettingsPage';
